refactor: Service 계층 도입
